Calculate if item is masterworked off its energy capacity in optimiser. Rename a few folders and files to be more consistent.
