+++ src/app/loadout-builder/mod-utils.ts
@@ -2,9 +2,9 @@ import { armor2PlugCategoryHashesByName } from 'app/search/d2-known-values';
 import { DestinyEnergyType } from 'bungie-api-ts/destiny2';
 import _ from 'lodash';
 import { DimItem } from '../inventory/item-types';
-import { mapArmor2ModToProcessMod, mapDimItemToProcessItem } from './processWorker/mappers';
-import { canTakeSlotIndependantMods, generateModPermutations } from './processWorker/processUtils';
-import { ProcessItem } from './processWorker/types';
+import { canTakeSlotIndependantMods, generateModPermutations } from './process-worker/processUtils';
+import { ProcessItem } from './process-worker/types';
+import { mapArmor2ModToProcessMod, mapDimItemToProcessItem } from './process/mappers';
 import {
   bucketsToCategories,
   knownModPlugCategoryHashes,
