+++ src/app/loadout-builder/process-worker/ProcessWorker.ts
