+++ src/app/loadout-builder/process-worker/processUtils.ts
