+++ src/app/loadout-builder/process-worker/set-tracker.ts
