+++ src/app/loadout-builder/process-worker/tsconfig.json
