+++ src/app/loadout-builder/utils.ts
@@ -2,7 +2,7 @@ import { DimItem, DimSocket } from 'app/inventory/item-types';
 import { DestinyEnergyType, TierType } from 'bungie-api-ts/destiny2';
 import { PlugCategoryHashes } from 'data/d2/generated-enums';
 import _ from 'lodash';
-import { ProcessItem } from './processWorker/types';
+import { ProcessItem } from './process-worker/types';
 import { LockedItemType, LockedMod, statValues } from './types';
 
 /**
